Change post-goal handling logic
The editor reported an error when the parent tries to remove multiple children in the same frame. However, as queueing removal makes it so that it only happens at the end of the frame, and the conditional to check if a new ball should be spawned happens at the end of the current frame, said conditional has been updated such that it will permit the spawning of a new ball if there are one or less balls in the tree.